requirements.txt
Django==4.2.11
djangorestframework==3.15.1
drf-yasg==1.21.7
python-dotenv==1.0.1